requirements.txt
alembic==1.16.4
asgiref==3.7.2
bcrypt==4.1.2
blinker==1.9.0
click==8.2.1
Flask==3.0.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
h11==0.16.0
itsdangerous==2.2.0
Jinja2==3.1.6
ledger-cli-toolkit==1.2.0
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
packaging==25.0
psycopg2-binary==2.9.9
PyJWT==2.10.1
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.42
tabulate==0.9.0
typing_extensions==4.14.1
uvicorn==0.24.0
Werkzeug==3.1.3
